+++ src/state/local-storage.js
@@ -1,0 +1,44 @@
+import { captureException } from '../../utils/sentry';
+
+const getStorage = () => {
+  try {
+    const storage = window.localStorage;
+    storage.setItem('test', 'test');
+    storage.getItem('test');
+    storage.removeItem('test');
+    return storage;
+  } catch (error) {
+    console.warn('Local storage not available, using memory store');
+  }
+  return null;
+};
+
+export const storage = getStorage();
+
+export const readFromStorage = (name) => {
+  if (storage) {
+    try {
+      const raw = storage.getItem(name);
+      if (raw !== null) {
+        return JSON.parse(raw);
+      }
+    } catch (error) {
+      captureException(error);
+    }
+  }
+  return undefined;
+};
+
+export const writeToStorage = (name, value) => {
+  if (storage) {
+    try {
+      if (value !== undefined) {
+        storage.setItem(name, JSON.stringify(value));
+      } else {
+        storage.removeItem(name);
+      }
+    } catch (error) {
+      captureException(error);
+    }
+  }
+};
